more work on recording videos...
proved that using ticks as approx for time is a bad idea.  On rpi, the cpu is not able to maintain 30fps. I am also worried that every animated component using time.time() will be slow.   I need to go back and add a single time.time() and pass the result down through all renderables :/